requirements.txt
kafka-python==0.9.4
retrying==1.2.3
msgpack-python==0.4.2
